Clean codes, minor fixies
Refine message browsing
Try move API services to submodule
